request home data

+++ app/src/main/java/com/glumes/androidkotlinsample/gankio/model/eyepetizer/HomeBean.kt
@@ -5,6 +5,6 @@ package com.glumes.androidkotlinsample.gankio.model.eyepetizer
  */
 data class HomeBean(val issueList: List<IssueListBean>,
                     val nextPageUrl: String,
-                    val nextPublishTime: Int,
+                    val nextPublishTime: Long,
                     val newestIssueType: String,
                     var dialog: Any?)
